requirements.txt
Appium-Python-Client==2.2.0;python_version >= '3.0'
pytest==7.1.2
allure-pytest==2.9.45